Adding authentication for embeddable gadgets
