+++ apps/portal/controllers/includes/dashboard-scripts.jag
@@ -24,14 +24,7 @@
     <%
         isAnonView = JSON.parse(request.getParameter('isAnonView'));
 
-        if(typeof isDefault == 'undefined')
-            isDefault = false;
-
-        // This is temporary change to avoid the anon view, when the user is not authenticated
-        if(isDefault){
-            viewType = 'default';
-        }
-        else if(isAnonView || !user) {
+        if(isAnonView || !user) {
             viewType = 'anon';
         } else {
             if(user && ((user.domain != superDomain && user.domain != urlDomain) ||
