+++ apps/portal/modules/dashboards.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -52,7 +52,7 @@ var registryUserPath = function (id, username) {
 
 /**
  * Get location where to store an OAuth application credentials
- * @param application name
+ * @param id
  * @returns string path
  */
 var registryOAuthApplicationPath = function (id) {
@@ -73,7 +73,7 @@ var getOAuthApplication = function (applicationId, callBack) {
     } else {
         callBack(null);
     }
-}
+};
 
 /**
  * Create an OAuth application against portal app
@@ -96,7 +96,7 @@ var createOAuthApplication = function (applicationId, clientCredentials) {
     } catch (exception) {
         throw "Error occurred while creating an OAuth application, " + exception;
     }
-}
+};
 
 /**
  * Finds a dashboard by its ID.
@@ -281,6 +281,59 @@ var allowed = function (dashboard, permission) {
     }
     if (permission.view) {
         return utils.allowed(user.roles, permissions.viewers);
+    }
+};
+
+/**
+ * Find a particular page within a dashboard
+ * @param {Object} dashboard Dashboard object
+ * @param {String} id Page id
+ * @return {Object} Page object
+ */
+var findPage = function (dashboard, id) {
+    var i;
+    var page;
+    var pages = dashboard.pages;
+    var length = pages.length;
+    for (i = 0; i < length; i++) {
+        page = pages[i];
+        if (page.id === id) {
+            return page;
+        }
+    }
+};
+
+/**
+ * Find a given component in the current page
+ * @param {Number} id
+ * @returns {Object}
+ * @private
+ */
+var findComponent = function (id, page) {
+    var i;
+    var length;
+    var area;
+    var component;
+    var components;
+    var type;
+
+    if ((user.domain != superDomain && user.domain != urlDomain) ||
+        (urlDomain && user.domain == superDomain && urlDomain != superDomain)) {
+        type = 'anon';
+    }
+
+    var content = (type === 'anon' ? page.content.anon : page.content.default)
+    for (area in content) {
+        if (content.hasOwnProperty(area)) {
+            components = content[area];
+            length = components.length;
+            for (i = 0; i < length; i++) {
+                component = components[i];
+                if (component.id === id) {
+                    return component;
+                }
+            }
+        }
     }
 };
